feat(frontend/adminDashboard): Implement Admin Dashboard and Employee Management
- Removed the old roadmap page and replaced it with a new structure.
- Added Admin controller for managing admin functionalities.
- Created Auth controller for user authentication and login handling.
- Developed Admin Dashboard view with statistics and recent inquiries.
- Implemented Employee Creation view for adding new employees.
- Created login page for admin access.
- Established a mood board page showcasing the visual identity of Batis Point.
- Updated roadmap page to reflect upcoming features and enhancements.
