+++ backend/app/Controllers/Users.php
@@ -7,83 +7,18 @@ use CodeIgniter\HTTP\ResponseInterface;
 
 class Users extends BaseController
 {
-    public function index()
+    public function showLandingPage()
     {
         return view('user/landingPage');
     }
 
-    public function login()
+    public function showMoodboard()
     {
-        return view('user/loginPage');
-        $session = session();
-        // Here i created rules for email and password
-        $validation = \Config\Services::validation();
-        // Variable comes from the html the id from the input
-        // Format: variable, human readable name, rules seperated by |
-        // So this following rule means variable email is Email which means it should not be null and has valid email format
-        $validation->setRule('email', 'Email', 'required|valid_email');
-
-        // Password shouldn't be null, minimum password length is 8 and maximum is 64.
-        $validation->setRule('password', 'Password', 'required', min_length(8), max_length(64));
-        $post = $request->getPost();
-        if (! $validation->run($post)) {
-            $session->setFlashdata('errors', $validation->getErrors());
-            $session->setFlashdata('old', $post);
-            return redirect()->back()->withInput();
-        }
-        $email = $request->getPost('email');
-        $userModel = new \App\Models\UsersModel();
-        $user = $userModel->where('email', $email)->first();
-        if (! $user) {
-            $session->setFlashdata('errors', ['email' => 'No account found for that email']);
-            $session->setFlashdata('old', ['email' => $email]);
-            return redirect()->back()->withInput();
-        }
-        $userArr = is_array($user) ? $user : (method_exists($user, 'toArray') ? $user->toArray() : (array) $user);
-        if (! password_verify($request->getPost('password'), $userArr['password_hash'] ?? '')) {
-            $session->setFlashdata('errors', ['password' => 'Incorrect password']);
-            $session->setFlashdata('old', ['email' => $email]);
-            return redirect()->back()->withInput();
-        }
-        $session->set('user', [
-            'id' => $userArr['id'] ?? null,
-            'email' => $userArr['email'] ?? null,
-            'first_name' => $userArr['first_name'] ?? null,
-            'last_name' => $userArr['last_name'] ?? null,
-            'type' => $userArr['type'] ?? 'client',
-            'display_name' => trim(($userArr['first_name'][0] ?? '') . ' ' . ($userArr['middle_name'][0] ?? '') . ' ' . ($userArr['last_name'] ?? '')),
-        ]);
-        $type = strtolower($userArr['type'] ?? 'client');
-        if ($type === 'manager') {
-            return redirect()->to('/admin/dashboard');
-        }
-
-        if ($type === 'client') {
-            return redirect()->to('/');
-        }
+        return view('user/moodboard');
     }
 
-    public function moodboard()
+    public function showRoadmap()
     {
-        return view('user/moodboardPage');
-    }
-
-    public function roadmap()
-    {
-        return view('user/roadmapPage');
-    }
-
-    public function employeeCreation()
-    {
-        return view('user/employeeCreationPage');
-    }
-    public function logout()
-    {
-        session()->destroy();
-
-        $params = session_get_cookie_params();
-        setcookie(session_name(), '', time() - 3600, $params['path'] ?? '/', $params['domain'] ?? '', isset($_SERVER['HTTPS']), true);
-
-        return redirect()->to('/');
+        return view('user/roadmap');
     }
 }
