+++ backend/app/Database/Seeds/UsersSeeder.php
@@ -22,6 +22,7 @@ class UsersSeeder extends Seeder
                 'email' => 'justinkirkendozo@gmail.com',
                 'password_hash' => 'Kulasa17',
                 'account_status' => '1',
+                'type' => 'manager',
                 'gender' => 'Male',
                 'created_at' => $now,
                 'updated_at' => $now,
@@ -33,6 +34,7 @@ class UsersSeeder extends Seeder
                 'email' => 'kulasaendozo@gmail.com',
                 'password_hash' => 'KulasaXD',
                 'account_status' => '1',
+                'type' => 'client',
                 'gender' => 'Female',
                 'created_at' => $now,
                 'updated_at' => $now,
