+++ backend/app/Views/components/footer.php
@@ -10,11 +10,11 @@
 
         <!-- Navigation Buttons -->
         <div class="flex justify-center md:justify-center gap-4">
-            <a href="<?= ('/moodboardPage'); ?>"
+            <a href="<?= ('/moodboard'); ?>"
                 class="px-5 py-2 rounded-full bg-[#F1B24A] text-[#1F3D2A] font-poppins font-medium hover:bg-[#d99c3d] transition duration-200 shadow-md">
                 Moodboard
             </a>
-            <a href="<?= ('/roadmapPage'); ?>"
+            <a href="<?= ('/roadmap'); ?>"
                 class="px-5 py-2 rounded-full border border-[#9EC590] text-[#FCFFF1] font-poppins font-medium hover:bg-[#9EC590]/20 transition duration-200">
                 Roadmap
             </a>
